refactor: Localize single-consumer constants (issue #643 phase 2)
Move constants from src/const.ts to their sole consumer:
- ACTOR_CACHE_*, APIFY_DOCS_CACHE_*, MCP_SERVER_CACHE_* -> src/state.ts
- USER_CACHE_MAX_SIZE/TTL_SECS -> src/utils/userid_cache.ts
- MCP_STREAMABLE_ENDPOINT -> src/mcp/const.ts (kept exported, used in tests)
- PROGRESS_NOTIFICATION_INTERVAL_MS -> src/utils/progress.ts (kept exported, used in tests)
- USER_AGENT_ORIGIN -> src/apify_client.ts
- SEGMENT_FLUSH_AT_EVENTS, SEGMENT_FLUSH_INTERVAL_MS -> src/telemetry.ts

Note: spec also listed ACTOR_SEARCH_ABOVE_LIMIT, but it has already been
localized to src/utils/actor_search.ts (no longer in src/const.ts).
