+++ src/state.ts
@@ -1,13 +1,12 @@
-import {
-    ACTOR_CACHE_MAX_SIZE,
-    ACTOR_CACHE_TTL_SECS,
-    APIFY_DOCS_CACHE_MAX_SIZE,
-    APIFY_DOCS_CACHE_TTL_SECS,
-    MCP_SERVER_CACHE_MAX_SIZE,
-    MCP_SERVER_CACHE_TTL_SECS,
-} from './const.js';
 import type { ActorDefinitionWithInfo, ApifyDocsSearchResult } from './types.js';
 import { TTLLRUCache } from './utils/ttl_lru.js';
+
+const ACTOR_CACHE_MAX_SIZE = 500;
+const ACTOR_CACHE_TTL_SECS = 30 * 60; // 30 minutes
+const APIFY_DOCS_CACHE_MAX_SIZE = 500;
+const APIFY_DOCS_CACHE_TTL_SECS = 60 * 60; // 1 hour
+const MCP_SERVER_CACHE_MAX_SIZE = 500;
+const MCP_SERVER_CACHE_TTL_SECS = 30 * 60; // 30 minutes
 
 export const actorDefinitionPrunedCache = new TTLLRUCache<ActorDefinitionWithInfo>(ACTOR_CACHE_MAX_SIZE, ACTOR_CACHE_TTL_SECS);
 export const searchApifyDocsCache = new TTLLRUCache<ApifyDocsSearchResult[]>(APIFY_DOCS_CACHE_MAX_SIZE, APIFY_DOCS_CACHE_TTL_SECS);
