+++ src/utils/progress.ts
@@ -1,7 +1,10 @@
 import type { ProgressNotification } from '@modelcontextprotocol/sdk/types.js';
 
 import type { ApifyClient } from '../apify_client.js';
-import { PROGRESS_NOTIFICATION_INTERVAL_MS, RELATED_TASK_META_KEY } from '../const.js';
+import { RELATED_TASK_META_KEY } from '../const.js';
+
+// The console uses const MIN_OBSERVER_INTERVAL_MILLIS = 3000 so it should be fine.
+export const PROGRESS_NOTIFICATION_INTERVAL_MS = 3000; // 3 seconds
 
 const TERMINAL_RUN_STATUSES = new Set(['SUCCEEDED', 'FAILED', 'ABORTED', 'TIMED-OUT']);
