+++ tests/unit/utils.progress.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, vi } from 'vitest';
 
-import { PROGRESS_NOTIFICATION_INTERVAL_MS, RELATED_TASK_META_KEY } from '../../src/const.js';
-import { createProgressTracker, formatRunStatusMessage, ProgressTracker } from '../../src/utils/progress.js';
+import { RELATED_TASK_META_KEY } from '../../src/const.js';
+import { createProgressTracker, formatRunStatusMessage, PROGRESS_NOTIFICATION_INTERVAL_MS, ProgressTracker } from '../../src/utils/progress.js';
 
 describe('ProgressTracker', () => {
     it('should send progress notifications correctly', async () => {
